Final code optimization for news-aggregator-app
