+++ Dockerfile
@@ -4,6 +4,8 @@ WORKDIR /app
 
 COPY package*.json ./
 
+COPY .env .env
+
 RUN npm install
 
 COPY . .
