+++ src/App.jsx
@@ -1,17 +1,21 @@
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
-import Header from './components/Header/Header'
-import NewsFeed from "./components/NewsFeed/NewsFeed";
-import NewsDetails from './components/NewsDetails/NewsDetails';
+import Header from './components/Header'
+import React, { Suspense } from "react";
+
+const NewsFeed = React.lazy(() => import("./page/NewsFeed"));
+const NewsDetails = React.lazy(() => import("./page/NewsDetails"));
 
 function App() {
 
   return (
     <Router>
       <Header/>
-      <Routes>
-        <Route path="/" element={<NewsFeed />} />
-        <Route path="/news/:id" element={<NewsDetails />} />
-      </Routes>
+      <Suspense fallback={<div>Loading...</div>}>
+        <Routes>
+          <Route path="/" element={<NewsFeed />} />
+          <Route path="/news/:id" element={<NewsDetails />} />
+        </Routes>
+      </Suspense>
     </Router>
   );
 }
