--- src/components/Footer/Footer.jsx
@@ -1,9 +1,0 @@
-
-
-const Footer = () => {
-  return (
-    <div>Footer</div>
-  )
-}
-
-export default Footer
