--- src/context/NewsContext.js
@@ -1,9 +1,0 @@
-
-
-const NewsContext = () => {
-  return (
-    <div>NewsContext</div>
-  )
-}
-
-export default NewsContext
