--- src/page/HomePage.js
@@ -1,9 +1,0 @@
-
-
-const HomePage = () => {
-  return (
-    <div>HomePage</div>
-  )
-}
-
-export default HomePage
