--- src/services/fetchArticles.js
@@ -1,9 +1,0 @@
-
-
-const fetchArticles = () => {
-  return (
-    <div>fetchArticles</div>
-  )
-}
-
-export default fetchArticles
